+++ src/components/GameBoard/ActionArea/EndGame/EndGame.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import Button from './Button';
+import Button from '../../../Base/Button';
 
 const renderGoodWin = () => {
-  return <div>Good wins! Arthur and Goodness prevails!</div>
+  return <div>Good wins! Arthur and Goodness prevail!</div>
 }
 
 const renderEvilWin = () => {
-  return <div>Bad wins! Mordred's dark forces of evil triumphs!</div>
+  return <div>Bad wins! Mordred's dark forces of Evil triumph!</div>
 }
 
 function EndGame (props) {
@@ -14,7 +14,7 @@ function EndGame (props) {
     <div style={{display: 'flex', flexDirection: 'column'}}>
       {props.status === 'GOOD_WIN' && renderGoodWin()}
       {props.status === 'EVIL_WIN' && renderEvilWin()}
-      <div style={{width: '2rems'}}><Button onClick={window.exitGame}>Exit Game</Button></div>
+      <div style={{width: '2rem'}}><Button onClick={window.exitGame}>Exit Game</Button></div>
     </div>
   )
 }
