+++ src/components/NotInGame/CreateForm.jsx
@@ -8,8 +8,8 @@ import Checkbox from '@material-ui/core/Checkbox';
 import Select from '@material-ui/core/Select';
 import MenuItem from '@material-ui/core/MenuItem';
 import TextField from "@material-ui/core/TextField";
-import Button from './Button';
-import { createRoom } from '../ApiUtils';
+import Button from '../Base/Button';
+import { createRoom } from '../../ApiUtils';
 
 export default class CreateForm extends React.Component {
   constructor(props) {
