+++ src/components/NotInGame/WaitingArea.jsx
